requirements.txt
Flask
PyMuPDF
Pillow
pytesseract
python-dotenv
fpdf2
openpyxl
gunicorn